+++ src/components/SearchBar/SearchBar.module.scss
@@ -1,0 +1,9 @@
+.modal_button{
+  background: none;
+  border: none;
+
+  padding: 1rem 2rem;
+  display: inline-block;
+  margin-right: 65px;
+  color: #99a1b3;
+}
