+++ src/models/SendMessageData.ts
@@ -1,0 +1,6 @@
+export interface sendMessageData {
+  to?: string;
+  cc?: string;
+  subject?: string;
+  messageText?: string;
+}
